+++ frameworks/qwik/src/index.ts
@@ -1,4 +1,16 @@
-export type { SubmitHandler } from '@formisch/core/qwik';
+export type {
+  DeepPartial,
+  FieldElement,
+  FormConfig,
+  PartialValues,
+  PathValue,
+  RequiredPath,
+  Schema,
+  SubmitHandler,
+  ValidArrayPath,
+  ValidationMode,
+  ValidPath,
+} from '@formisch/core/qwik';
 export * from '@formisch/methods/qwik';
 export * from './components/index.ts';
 export * from './hooks/index.ts';