+++ frameworks/svelte/src/index.ts
@@ -1,4 +1,16 @@
-export type { SubmitHandler } from '@formisch/core/svelte';
+export type {
+  DeepPartial,
+  FieldElement,
+  FormConfig,
+  PartialValues,
+  PathValue,
+  RequiredPath,
+  Schema,
+  SubmitHandler,
+  ValidArrayPath,
+  ValidationMode,
+  ValidPath,
+} from '@formisch/core/svelte';
 export * from '@formisch/methods/svelte';
 export * from './components/index.ts';
 export * from './runes/index.ts';